Created seperate view files
